[UPDATE] update payment in & cart

+++ Models/PaymentsModel.cs
@@ -7,7 +7,7 @@ namespace Artfusion.Models
         [Key]
         public string? PaymentId { get; set; }
         public string? UserId { get; set; }
-        public double Amount { get; set; }
+        public long Amount { get; set; }
         public DateTime? CreatedAt { get; set; }
     }
 }
